work on elco data display

+++ mcare-dashboard-web/src/main/java/org/mcare/common/interfaces/DatabaseRepository.java
@@ -5,10 +5,12 @@ import java.util.List;
 public interface DatabaseRepository {
 	
 	public <T> long save(T t);
+	
 	public <T> int update(T t);
+	
 	public <T> boolean delete(T t);
 	
-	public <T> T findById(long id, String fieldName, Class<?> className);
+	public <T> T findById(int id, String fieldName, Class<?> className);
 	
 	public <T> T findByKey(String value, String fieldName, Class<?> className);
